consensus separated block building
