+++ cmd/erigon-el-mock/server.go
@@ -14,7 +14,7 @@ import (
 	"github.com/ledgerwatch/erigon-lib/gointerfaces/execution"
 	types2 "github.com/ledgerwatch/erigon-lib/gointerfaces/types"
 	"github.com/ledgerwatch/erigon-lib/kv"
-	"github.com/ledgerwatch/erigon/turbo/engineapi"
+	"github.com/ledgerwatch/erigon/turbo/engineapi/engine_types"
 	"github.com/ledgerwatch/erigon/turbo/services"
 
 	"github.com/ledgerwatch/erigon/core/rawdb"
@@ -201,7 +201,7 @@ func (e *Eth1Execution) GetBody(ctx context.Context, req *execution.GetSegmentRe
 	if err != nil {
 		return nil, err
 	}
-	rpcWithdrawals := engineapi.ConvertWithdrawalsToRpc(body.Withdrawals)
+	rpcWithdrawals := engine_types.ConvertWithdrawalsToRpc(body.Withdrawals)
 	unclesRpc := make([]*execution.Header, 0, len(body.Uncles))
 	for _, uncle := range body.Uncles {
 		unclesRpc = append(unclesRpc, HeaderToHeaderRPC(uncle))
